requirements.txt
click>=8.0.0
colorama>=0.4.4
rich>=13.0.0
openai>=1.0.0
anthropic>=0.8.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0
jsonschema>=4.0.0
ast-grep-py>=0.12.0
pathspec>=0.11.0
tabulate>=0.9.0
